+++ Android/TogeDuck/app/src/main/java/com/idle/togeduck/quest/exchange/model/ExchangeService.kt
@@ -17,20 +17,20 @@ import retrofit2.http.Query
 interface ExchangeService {
     @GET("events/{event_id}/trades")
     suspend fun getExchangeList(
-        @Path("event_id") eventId: Int,
+        @Path("event_id") eventId: Long,
         @Query("page") page: Int,
         @Query("size") size: Int
     ) : Response<ExchangeListResponse>
 
     @GET("events/{event_id}/trades/{trade_id}")
     suspend fun getExchangeDetail(
-        @Path("event_id") eventId: Int,
+        @Path("event_id") eventId: Long,
         @Path("trade_id") tradeId: Int,
     ) : Response<ExchangeDetailResponse>
 
     @GET("events/{event_id}/my-trades")
     suspend fun getMyExchangeList(
-        @Path("event_id") eventId: Int,
+        @Path("event_id") eventId: Long,
         @Query("page") page: Int,
         @Query("size") size: Int
     ) : Response<ExchangeListResponse>
@@ -38,7 +38,7 @@ interface ExchangeService {
     @Multipart
     @POST("events/{event_id}/trades")
     suspend fun postExchange(
-        @Path("event_id") eventId: Int,
+        @Path("event_id") eventId: Long,
         @Part image: MultipartBody.Part,
         @Part content: String,
         @Part duration: Int
@@ -47,7 +47,7 @@ interface ExchangeService {
     @Multipart
     @PATCH("events/{event_id}/trades/{trade_id}")
     suspend fun updateExchange(
-        @Path("event_id") eventId: Int,
+        @Path("event_id") eventId: Long,
         @Path("trade_id") tradeId: Int,
         @Part image: MultipartBody.Part,
         @Part content: String,
@@ -56,25 +56,25 @@ interface ExchangeService {
 
     @DELETE("events/{event_id}/trades/{trade_id}")
     suspend fun deleteExchange(
-        @Path("event_id") eventId: Int,
+        @Path("event_id") eventId: Long,
         @Path("trade_id") tradeId: Int,
     ) : Response<DefaultResponse>
 
     @POST("events/{event_id}/trades/{trade_id}/requests")
     suspend fun requestExchange(
-        @Path("event_id") eventId: Int,
+        @Path("event_id") eventId: Long,
         @Path("trade_id") tradeId: Int,
     ) : Response<DefaultResponse>
 
     @POST("events/{event_id}/trades/{trade_id}/requests/accept")
     suspend fun requestAcceptExchange(
-        @Path("event_id") eventId: Int,
+        @Path("event_id") eventId: Long,
         @Path("trade_id") tradeId: Int,
     ) : Response<DefaultResponse>
 
     @POST("events/{event_id}/trades/{trade_id}/requests/reject")
     suspend fun requestRejectExchange(
-        @Path("event_id") eventId: Int,
+        @Path("event_id") eventId: Long,
         @Path("trade_id") tradeId: Int,
     ) : Response<DefaultResponse>
 }
